requirements.txt
captum==0.6.0
joblib==1.3.2
matplotlib==3.8.3
numpy==1.26.4
openml==0.14.1
pandas==2.1.1
plotly==5.18.0
PyYAML==6.0.1
scikit-learn==1.3.0
scipy==1.11.3
seaborn==0.13.0
torch==2.0.1
tqdm==4.66.1
xgboost==2.0.0